[refactor/#155] 토글 훅을 훅 디렉토리로 이동
